Fold `util` module into `bash` and `fish`
