+++ src/bash.rs
@@ -187,7 +187,8 @@ impl Bash {
 // ----------------------------------------------------------------------------
 
 mod bytes {
-    use crate::{ascii::Char, util::u8_to_hex_escape};
+    use super::u8_to_hex_escape;
+    use crate::ascii::Char;
 
     pub enum Prepared {
         Empty,
@@ -239,7 +240,8 @@ mod bytes {
 // ----------------------------------------------------------------------------
 
 mod text {
-    use crate::{utf8::Char, util::u8_to_hex_escape};
+    use super::u8_to_hex_escape;
+    use crate::utf8::Char;
 
     pub enum Prepared {
         Empty,
@@ -286,5 +288,33 @@ mod text {
             }
         }
         sout.push(b'\'');
+    }
+}
+
+// ----------------------------------------------------------------------------
+
+/// Escape a byte as a 4-byte hex escape sequence.
+///
+/// The `\\xHH` format (backslash, a literal "x", two hex characters) is
+/// understood by many shells.
+#[inline]
+fn u8_to_hex_escape(ch: u8) -> [u8; 4] {
+    const HEX_DIGITS: &[u8] = b"0123456789ABCDEF";
+    [
+        b'\\',
+        b'x',
+        HEX_DIGITS[(ch >> 4) as usize],
+        HEX_DIGITS[(ch & 0xF) as usize],
+    ]
+}
+
+#[cfg(test)]
+#[test]
+fn test_u8_to_hex_escape() {
+    for ch in u8::MIN..=u8::MAX {
+        let expected = format!("\\x{ch:02X}");
+        let observed = u8_to_hex_escape(ch);
+        let observed = std::str::from_utf8(&observed).unwrap();
+        assert_eq!(observed, &expected);
     }
 }
